requirements.txt
opencv-python
pillow
requests
pynput
pygetwindow
